Remove Debugging
Also removed old unnecessary FB user functions (cannot be used anyhow)
Removal of old root icon
Result class renamed functions
Moved share button into Result
Result class can retrieve, delete and create a result rather than having
it be done elsewhere
Results vs Result mixups are now gone by removing the unnecessary
Results class and handling most of the functionality in the Result class
as part of a model and page.
facebook share result was embedded into a function and grabs the current
App ID from the SDK
Account now allows results to be hidden if they want to
The account page is also using ajax to present a better experience and
the icons were changed to font awesome as it was already included at the
time
Time now says 'just now' when under a minute rather than having to
update the times each second to half a minute but this only occurs on
the account page.
Result header was removed as it required too much
2nd person to 3rd person conversion now complete for public viewers